backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
transformers
torch
whisper
pydub
gtts